Split test module to reflect new structure
